% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_utils.R
\name{check_eigen_decomposition}
\alias{check_eigen_decomposition}
\title{check_eigen_decomposition}
\usage{
check_eigen_decomposition(e, tol = 1e-06)
}
\arguments{
\item{e}{the eigen decomposition to check}

\item{tol}{the threshold below which a number is said to be effectively zero, defaults to 1e-6}
}
\value{
The eigen values with any values with absolute value less than tol zeroed.
}
\description{
Grabbed from MASS.  Useful to sparsify matrices when some eigenvalues are essentially zero.
}
