% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariate_mat}
\alias{covariate_mat}
\title{Example covariate matrix}
\format{A matrix with 200 rows and 4 variables}
\usage{
covariate_mat
}
\description{
Example covariate matrix
}
\keyword{data}
