% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeScan.R
\docType{class}
\name{GenomeScan-class}
\alias{GenomeScan-class}
\alias{GenomeScan}
\title{GenomeScan}
\description{
A Reference Class implementing a Genome Scan
}
\section{Fields}{

\describe{
\item{\code{Data}}{Things the user inputs.
They have interpretable meaning and define the GenomeScan.
Currently: y, X, G, K, and weights.}

\item{\code{Intermediates_per_scan}}{Things the GenomeScan will compute once per scan.
They are mathetmatical tools that can't really be interpreted.
Currently: L, eigen_L, and LL_null.}

\item{\code{Intermediates_per_locus}}{Things the GenomeScan will compute once per locus.
They are mathetmatical tools that can't really be interpreted.
Currently: XG}

\item{\code{Intermediates_per_fit}}{Things the Genomecan will compute many times per locus (once per trial fit on that locus).
These are interpretable but rapidly changing and not guaranteed to be finalized or optimal.
Currently: M, LDV, and h2}

\item{\code{Results}}{The results of the GenomeScan.
Currently: The h2 that maximizes the LL at each locus and the LR as compared with the no-locus (null) model.}
}}


