% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeScan.R
\name{GenomeScan_calc_multiplier_eigen}
\alias{GenomeScan_calc_multiplier_eigen}
\title{calc_multiplier_eigen}
\value{
an object of class GenomeScan
}
\description{
Compute a multiplier (aka rotation) matrix.  Details in in h2lmm_math_RWC.Rmd.
}
