% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeScan.R
\name{GenomeScan_fit_locus_given_h2}
\alias{GenomeScan_fit_locus_given_h2}
\title{fit_locus_given_h2}
\value{
an object of class GenomeScan
}
\description{
Fit one locus at a specified value of h2.  Should not typically be called by a user.
}
