% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeScan.R
\name{GenomeScan_initialize}
\alias{GenomeScan_initialize}
\title{initialize}
\arguments{
\item{y}{vector of length n - the phenotype of each of n individuals}

\item{X}{matrix of dimension n-by-c - the covariate value of each individual for c covariates}

\item{G}{matrix of dimension n-by-p - the genotype of each individual at p loci}

\item{K}{matrix of dimension n-by-n - the covariance of the phenotype}

\item{w}{matrix of dimension n-by-n - the covariance of the phenotype}
}
\value{
an object of class GenomeScan
}
\description{
Initialize a GenomeScan
}
