% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locus_list}
\alias{locus_list}
\title{Example locus list}
\format{A list of matrices, where each matrix has 200 rows and 1 column}
\usage{
locus_list
}
\description{
Example locus list
}
\keyword{data}
