% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se_mean_per_strain}
\alias{se_mean_per_strain}
\title{Example standard error of the mean per strain}
\format{A vector of length 200}
\usage{
se_mean_per_strain
}
\description{
Example standard error of the mean per strain
}
\keyword{data}
